pyproject.toml
[tool.poetry]
name = "microbit"
version = "0.1.0"
description = ""
authors = ["Igor Brejc <igor.brejc@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
pseudo-microbit = "^1.10"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
